Thread safety improvements with `LockedBuffer` (#125)
* Initial LockedBuffer implementation

* Remove blosc.h include from compression params header file

* (wip) redefine Byte* to *<uint8_t>

* Finish LockedBuffer implementation

* Improve error logging and cleanup in filesystem integration tests

* Implement a method for determining which chunk indices belong in a given shard and chunk/shard layer

* Replace std::byte with uint8_t in Sink code and tests

* Use LockedBuffer in FrameQueue

* Use LockedBuffer in Downsampler

* Use LockedBuffer when writing to Nodes

* #include <cstdint> in definitions.hh for uint8_t

* Remove dead code in V3Array

* Remove some more dead code

* Try the OMP parfor without restricting threads

* Revert "Try the OMP parfor without restricting threads"

This reverts commit df65c9213bffa1b4fc968fa525eca5007d708e35.

* Extract lambdas

* (wip): don't push jobs to thread pool, remove OMP thread count restriction

* Filter out empty string when getting parent paths

* Revert "Filter out empty string when getting parent paths"

This reverts commit cbeec22b047f66f42d784410a7e9cdfe62d57081.

* Don't try to create empty directories

* (wip): fix broken control flow in V2Array::compress_and_flush_data_()

* (wip): fix broken control flow in V3Array::compress_and_flush_data_()

* Send `make_file_sinks` kernel to thread pool

* Send `make_dirs` kernel to thread pool

* Send
`V2Array::compress_and_flush_data_()` kernel to thread pool

* Send
`V2Array::compress_and_flush_data_()` kernels to thread pool

* Update error message on failure to send frame process job to thread pool

* (wip): can still deadlock with single thread

* Check number of threads when submitting a job to the thread pool so as not to deadlock

* Check for null Sink pointers

* Remove unused includes / static variable in array.cpp.

* Remove unused header

* Fix bug in Downsampler that downsamples original frame data in place

* Don't clear data on `FrameQueue::pop`

* (wip) let's see something real fast

* Shrink testing arrays to reduce memory pressure

* Try `python -m pybind11` first

* Add a debug test job with core dumps (Ubuntu only)

* Add a debug test job with core dumps (Ubuntu only)

* Make latch management less complicated

* Make latch management even less complicated

* Get crash dumps on Windows

* Get crash dumps on Windows

* Get core/crash dumps for Python files.

* The latch tweaks will continue until the crashes improve

* Remove test-debug.yml

* Revert "Remove test-debug.yml"

This reverts commit 3ad18e88960e35e92787be82bc1a8f5d8c84fbba.

* Add a lock to the FrameQueue methods push/pop

* Remove test-debug.yml

* Reinstate logging macros

* Comment decision to enforce a minimum of 2 threads here possible.

* Try finding Python before finding pybind11

* Whitespace

+++ tests/integration/stream-zarr-v3-3d-multiscale-to-filesystem.cpp
@@ -396,7 +396,13 @@ verify_file_data(int level)
                                x_file.string(),
                                "'");
                         const auto file_size = fs::file_size(x_file);
-                        EXPECT_LT(size_t, file_size, expected_file_size);
+                        EXPECT(file_size < expected_file_size,
+                               "Expected file size < ",
+                               expected_file_size,
+                               " for file ",
+                               x_file.string(),
+                               ", got ",
+                               file_size);
                     }
 
                     CHECK(!fs::is_regular_file(
@@ -478,12 +484,14 @@ main()
 
         verify();
 
-        // Clean up
-        fs::remove_all(test_path);
-
         retval = 0;
     } catch (const std::exception& e) {
         LOG_ERROR("Caught exception: ", e.what());
+    }
+
+    // cleanup
+    if (fs::exists(test_path)) {
+        fs::remove_all(test_path);
     }
 
     return retval;
